Allow mssf-core user to not depend on windows crate, and merge 2 test apps into 1 (#26)
We have 2 sample apps for stateless service, and each tests some part of
the mssf-core and com wrapping code. We only run echomain app in the CI,
so some of the test paths were not executed.
Merging 2 apps in a single one so that CI can cover all the test paths.

Also re-exported some times from windows crate in mssf-core, so that
users of this crate does not have to explicitly depend of windows crate.