Add Tokio and config-rs features (#37)
As discussed offline, I have some oddball integration scenarios where
much of the SF runtime interaction is already handled by existing
application logic written in other languages, but we'd like to move
parts of the system into Rust.

At least for now, my project is not making use of any of the
functionality in mssf-core that actually necessitates use of Tokio, nor
are we currently using the config_rs crate. At the same time, mssf_com
is a bit too low level to be ideal for us - we have no desire to
duplicate / rewrite existing Rust wrappers such as:
* mssf_core::runtime::config types
* mssf_core::runtime::ActivationContext

To enable minimizing dependencies and compile times for those of use who
don't need the entire runtime at present, this PR adds two default-on
features (naming suggestions of course welcome) - one for config-rs, one
for tokio (tokio feature also requires ctrlc as that's small and part of
the Executor, and I'm assuming most native-Rust applications that are ok
with tokio will want that too).

It also moves some of the code form runtimes/mod.rs to its own file to
make the conditional complication a bit less messy, though it could be
cleaner.

Additionally, one of the echomain (echomain-stateful) examples is
adjusted to demonstrate that config_rs feature is optional. The example
already did not need any of the APIs that are now gated behind that
feature, demonstrating that a thinner version of mssf_core could have
utility to pure Rust SF applications as well.
I think it's just the hostname function that even needs the tokio_async
feature, but haven't tried to create a "neither feature" example yet.