Remove generated client wrapper and generator (#44)
Generating a safe wrapping for fabric client is too ambitious. In
practice the generated wrapper needs to be hand wrapped anyway in
practice as observed in the work of doing FabricClient safe wrapping.
The complete safe wrapping in practice is never generated.
See: https://github.com/Azure/service-fabric-rs/pull/19

The generator is a pain to maintain as well due to windows-metadata does
not have an stable api yet. Every time a windows-rs upgrade, the
generator requires a lot of fixes.

For all these reasons, there is no advantage maintaining the generated
safe wrappings, so I am removing it in this PR.
In future all safe wrapping are going to be manually wrapped, as they
already are. If the windows-metadata or windows-bindgen crate supports
api to access to the unsafe com bindings syntax structures in future,
this work can be revived.