[FabricClient] Support 'register service notification filter' api (#60)
Partial impl for #57 
FabricClient can register the service notification filter and unregister
it.
See dotnet
[documentation](https://learn.microsoft.com/en-us/dotnet/api/system.fabric.fabricclient.servicemanagementclient.registerservicenotificationfilterasync?view=azure-dotnet)
The registration mechanism is used inside the dotnet implementation of
[servicepartitionresolver](https://learn.microsoft.com/en-us/dotnet/api/microsoft.servicefabric.services.client.servicepartitionresolver?view=azure-dotnet)

The registered service to get notification will no longer need to pass
the previous endpoint resolve result to force a refresh. This is tested
in the stateful example app. But the notification seems to be not very
fast (lag of a couple of seconds). But this notification utilizes
FabricClient cache, and reduces the cluster api load.
I will compare with direct resolution next.