[sync] Cancellation waits for SF task to complete (#62)
Currently when cancelling the SF operation proxy operation, the rust
future is short circuited when the Cancellation Token is cancelled. This
means that rust code does not wait for SF operation to complete or react
to the cancel signal.
This PR changes the behavior to let SF operation proxy wait for sf
operation to complete. If the operation is cancelled, wait for sf
operation's response. It is sf operation's responsibility to react to
the signal, and complete faster (with and error code).
If user wants to discard SF task result and short circuit without
waiting, user needs to use select! of the sf operation and the
cancellation token (not recommended or yet tested).