Add memory barrier for bridge context (#99)
mssf IFabricAsyncOperationContext implementation BridgeContext3 requires
memory barrier to synchronize the operation complete status and the ctx
content.
For example, a problematic way to use SF COM API:
Caller of SF COM API can call Begin operation, poll on ctx's status in a
loop until complete using IsCompleted() api , and End operation without
barriers. Without this change, this might cause the caller to forever be
looping and polling the bool completed status. This is because the
background thread set the status and result content in some memory
location, but the memory location is not synchronized to the caller's
thread.
There might be other ways including using callbacks that can result the
same error behavior.

This PR fixes this to add proper memory barrier to ensure that completed
status and the async operation result are visible by the callers'
threads.