Add small benchmark for SF bridge perf (#104)
SF COM C api and rust code are connected by bridge and proxy rust
wrappers.
This PR adds a simple benchmark to compare an async operation
performance when there is no wrapping, 1 level of wrapping, and 2 levels
of wrapping. The test shows that there is barely any degradation when
wrapping is inserted.
example output:
```
---- sync::cancel::test::small_bench_test stdout ----
count0: 47629, count1: 47628, count2: 47628
```
This shows that the wrapper code does not affect much of the async
operation performance.