Change safe APIs to return mssf Error (#131)
mssf Error can be converted to ErrorCode enum easily and user app can
check the enum to determine why API failed.
Raw api returns windows_core::Error and typically user converts to mssf
Error all the time after mssf API call. This PR changes the return time
of all safe APIs to mssf Error so it is easy to handle failures, and it
displays human readable messages.
Renamed FabricError to Error and this is the mssf Error we will use.
Renamed FabricErrorCode to ErrorCode since it is clear.

One part of this change that requires careful review is the Bridge layer
code where the ctx from Begin and End functions are not type checked but
casted. I checked them manually.