Start adding support for FabricClientSettings and FabricSecurityCredentials (#139)
I'd like to have support for ergonomic (i.e. safe) versions of
FABRIC_CLIENT_SETTINGS and FABRIC_SECURITY_CREDENTIALS
This makes necessary error handling changes (that are breaking), as
FabricClientBuilder::build() now can error if there are invalid
settings. It's a small step towards #140
* For now, the error type remains opaque (as it isn't actually
instantiated anyway).
* The FabricClientSettings struct remains empty (and adding new Optional
defaults shouldn't breaking)
* The FabricSecurityCredentials enum is also empty (and marked
non_exhaustive, so adding more entries in future will not be a breaking
change)
* Also add some trivial wrapper types for the next PR, empty modules,
and so on.