Enable info span for bridge functions (#148)
Original instrument creates a debug level span that has no effect on
traces of info level.
```rs
 tracing::instrument(skip_all, level = "debug", ret, err)
```
Changed the span level to be info by default. But trace the ret value
only in debug trace.
```rs
 tracing::instrument(skip_all, ret(level = "debug"), err)
```
The executor trace is fixed up to contain the correct span and panic
message.
For example
```log
025-04-09T21:35:26.812342Z ERROR BeginOpen: mssf_core::runtime::executor: DefaultJoinHandle: background task failed: E_UNEXPECTED, msg:task 18 panicked with message "test panic"
```