Add get_deployed_replica_detail to fabric client (#149)
**Description**
Currently, The GetDeployedReplicaListAsync is exposed in Fabric Client
API. This is required to find the read and write status of a Stateful
Replica along with other fields.
This PR adds the `get_deployed_replica_detail` API to fabric query
client. However, the list types are ignored for now as the memory owned
by com object must be saved when using list with shared pointer.
**Testing:**
Added unit test which invokes get_deployed_replica_detail and verifies
stateful server parition write and read status

---------

Co-authored-by: Hemanth Sandrana <sahemant@microsoft.com>