0.1.0 Release (#160)
Changed mssf crate versions to 0.1.0.
The crates are stable enough that we now should utilize patch version
for small fixes.
From now on the patch version of the mssf crates may differ, and each
crate may have different release cadence. For example, the mssf-pal
crate has not been changed for a long time, so we can skip publishing
the same code with new versions over and over again.