Mitigate test crash (#190)
FabricClient clean up on process exit causes seg fault. It is most
likely due to some dangling ptr leaked into the thread pool. There is
internally tracked bug which I will debug at another time, it is taking
too long.
Our rust code is not the problem. So this PR adds a 1 second delay for
the longest running test for the process exit, and it seems to be
reliable since the pipeline does not fail for many reruns.
Mitigate #184