Add CodePackageActivationContext::get_config_package_names and CodePackageActivationContext::get_data_package_names (#192)
I have a version of get_config_package_names in my repo; but it makes
more sense to have the library provide it, exaclty like
get_code_package_names. So add it.

I don't have immediate need for data package names, but I already
happened to see it in source while checking the implementation of
GetConfigPackageNames, and confirmed that it's not actually able to fail
today either, and I may want it eventually (though there's more work to
be done to expose Data Packages more fully if I ever do want it).