Remove WStringWrap type (#208)
WString supports direct FFI conversion with PCWSTR since the custom
implementation in mssf-pal.
WStringWrap originally was created to interoperate with widnows HSTRING,
and it is no longer needed.
Modified all WStringWrap usage to use WString directly.
Changed to use Uri type when the FFI type is FABRIC_URI.
Changed the signature of stateless and stateful factories to move the
WString to avoid user coping.