[Runtime] Use move instead of ref in traits (#209)
In user facing traits like StatefulServiceReplica, many params were
passing by ref, and user ends up copying the ref in the callback.
But the refs do not need to be preserved in the interop layer, so it can
be directly moved to user callback.
The reason it was implemented as ref in the first place was that SF FFI
api is using ref/ptr, and I just blindly followed the pattern.
This PR changes those refs to move and it can get rid of user coping in
most cases.