Add simple cmake module to pkg sample apps (#218)
Packaging sample apps in this repo has many repetitive cmake scripts.
This PR adds a simple cmake module/function to capture the packaging
logic. The packaging logic is very basic and does not cover general case
where an app can have multiple service types etc.

In the meantime, the CI on the main branch is broken due to linux
runners out of disk space for devcontainer jobs.
It is fixed by changing docker engine to use the data disk instead of
the os disk.

Also changed the dev-container user to be the same has the user on host
machine to avoid repo permissions problem on host.