[mock] Initial support for mock unit testing of Stateless Service (#226)
Added a MockDriver that can register the stateless service factory and
create/delete instances.
User can test their SF service easily in Rust unit test. This is a bare
minimum implementation, and the mock is subject to change in future.

To support mock testing, trait needs to be `Box<dyn T>` compatible, so
we switch back to our initial implementation using
[async_trait](https://crates.io/crates/async-trait). Refactored
IStatelessServicePartition into a trait for better mock testing.

The end goal is to be able to mock unit test a service+partition
creation and stateful service failover in unit test, and this PR is the
first step.