[core] Use Arc instead of Box for Partition object (#230)
dyn IStatefulServicePartition and dyn IStatelessServicePartition trait
implementations in practice always require cloning in applications. So,
it is better to change it to Arc from Box, so that application can avoid
wrap another layer of Arc. And the trait does not have to implement
clone_box() function to implement boxed clone.